LayoutTests/platform/win/accessibility/aria-tab-roles-expected.txt

Crust
Veges
Select Crust

This tests that the aria roles for tab, tabpanel and tablist work as expected correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


tabList.role = page tab list
tab1.role = list item
tab1.title = Crust
PASS tab1.childrenCount is 0
tab2.role = list item
tab2.title = Veges
tabPanel.role = grouping
tabPanel.subrole = 
PASS successfullyParsed is true

TEST COMPLETE
